Migrate to .NET9 and Read SaveID
