+++ SOURCE/.editorconfig
@@ -1,0 +1,4 @@
+﻿[*.cs]
+
+# CA1416: Validate platform compatibility
+dotnet_diagnostic.CA1416.severity = none
