--- SOURCE/BG-Save-Backup/Classes/StringExtensions.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BG3_Save_Backup.Classes {
-    public static class StringExtensions {
-        public static string NullIfEmpty(this string value) {
-            return string.IsNullOrEmpty(value) ? null : value;
-        }
-        public static string NullIfWhiteSpace(this string value) {
-            return string.IsNullOrWhiteSpace(value) ? null : value;
-        }
-    }
-}
