--- SOURCE/BG-Save-Backup/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft-WindowsAPICodePack-Core" version="1.1.5" targetFramework="net481" />
-  <package id="Microsoft-WindowsAPICodePack-Shell" version="1.1.5" targetFramework="net481" />
-</packages>
