+++ SOURCE/BG3 Save Backup.sln
@@ -1,20 +1,37 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29318.209
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35521.163
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BG3 Save Backup", "BG-Save-Backup\BG3 Save Backup.csproj", "{8E9E0A6D-E687-4D5E-B698-D8B97907CCD1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BG3SaveBackup", "BG3SaveBackup\BG3SaveBackup.csproj", "{5B2453A9-634A-4DB4-AF65-11FE8C16D632}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{2FBE795D-040E-451E-9A01-304768FF7B15}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8E9E0A6D-E687-4D5E-B698-D8B97907CCD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8E9E0A6D-E687-4D5E-B698-D8B97907CCD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8E9E0A6D-E687-4D5E-B698-D8B97907CCD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8E9E0A6D-E687-4D5E-B698-D8B97907CCD1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Debug|x64.Build.0 = Debug|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Debug|x86.Build.0 = Debug|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Release|x64.ActiveCfg = Release|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Release|x64.Build.0 = Release|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Release|x86.ActiveCfg = Release|Any CPU
+		{5B2453A9-634A-4DB4-AF65-11FE8C16D632}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
