+++ SOURCE/BG3SaveBackup/App.config
@@ -2,14 +2,11 @@
 <configuration>
     <configSections>
         <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-            <section name="BG3_Save_Backup.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+            <section name="BG3SaveBackup.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
         </sectionGroup>
     </configSections>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8.1" />
-    </startup>
     <userSettings>
-        <BG3_Save_Backup.Properties.Settings>
+        <BG3SaveBackup.Properties.Settings>
             <setting name="LarianSaveLoc" serializeAs="String">
                 <value />
             </setting>
@@ -20,8 +17,8 @@
                 <value>True</value>
             </setting>
             <setting name="HonorOnly" serializeAs="String">
-                <value>False</value>
+                <value>True</value>
             </setting>
-        </BG3_Save_Backup.Properties.Settings>
+        </BG3SaveBackup.Properties.Settings>
     </userSettings>
 </configuration>
