+++ SOURCE/BG3SaveBackup/GlobalUsings.cs
@@ -1,0 +1,3 @@
+﻿global using BG3SaveBackup.Classes;
+global using BG3SaveBackup.Properties;
+global using BG3SaveBackup.Forms;
