+++ SOURCE/BG3SaveBackup/Images/DefaultIcon.ico
